Enhance and test support for PostgreSQL (#189)
* wip

* wip

* Update OrderByTest.php

* Fixed hasOneOfMany for MySQL and Postgres

* Removed readme

* Handling left joins in one of many

* Add phpunit.xml to gitignore

* Remove support for unsupported PHP/Laravel versions

* Remove DB setup in test case

* Add CI jobs for testing the package across multiple DBs

* Update DB connection credentials

* Update ci.yaml

* Ensure correct DB name across Laravel versions

* Set DB var in action instead

* Preparing to run tests on all dbs

* Fixed test

* Update JoinRelationshipTest.php

* Added additional test

* Fixed powerJoinHas with hasOneOfMany

* Fixed concat checks

* Fix join instance counting on MySQL

* Update TestCase.php

* Fixed OrderByTest::test_order_by_relationship_min_and_max in MySQL

* Update TestCase.php

* Update RelationshipsExtraMethods.php

* Fix Postgres tests again

* Added custom one of many logic for MySQL

* Removed old Postgres check

* Update JoinRelationshipTest.php

* Update 2020_03_16_000000_create_tables.php

* Dropped PowerJoins trait

* Removed PowerJoins traits

---------

+++ tests/database/migrations/2020_03_16_000000_create_tables.php
@@ -32,7 +32,7 @@ class CreateTables extends Migration
         Schema::create('groups', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->tinyInteger('access_level')->default(0);
+            $table->integer('access_level')->default(0);
             $table->timestamps();
         });
